chore: move custom component test to showcase example
