+++ examples/showcase/components/CustomComponent.vue
