--- packages/storybook-nuxt/playground/components/PreHello.vue
@@ -1,6 +1,0 @@
-<template>
-  <pre>
-        <h1> Hello </h1>
-        <NuxtLink to="https://github.com/storybook-vue/storybook-nuxt" target="_blank">Storybook-Nuxt World</NuxtLink>
-    </pre>
-</template>
